question_generator/requirements.txt
duckduckgo-search>=5.3.1
curl_cffi>=0.7.1
beautifulsoup4>=4.12.3
openai>=1.14.0
pydantic>=2.6.4
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
